Wildfire risk prediction in Galicia
An unsupervised density-based clustering technique is used to predict the risk of wildfire, having as use case the fires of Ponteareas, Galicia in 2017. Code was written following the method proposed in Salehi et al 'Dynamic and Robust Wildﬁre Risk Prediction System: An Unsupervised Approach' IBM Research paper[1]. The model achieved an AUC of 0.995 compared to the current way of estimating fire risk, the Forest Fire Danger Index (FFDI) which obtained an AUC of 0.69. Looking at the Precision-Recall curve (more appropriate for rare events) the obtained Average Precision-Recall score was 0.94, far better than the 0.23 score for the FFDI. 

However, these metrics need to be taken into account with caution. That a wildfire occurred one day doesn't mean that there wasn't any risk the day before (in fact there was, as predicted by our model), but to compute the metrics only the days with fire were considered positives. In any case, The proposed model outperformed the current way of estimating fire risk. If used together with the HIRLAM 0.05 model to obtain weather data on every location (not only in places near weather stations as we did) in real time it could be a powerful tool to combat wildfires in Galicia, although of course more tests would be needed.

[1] Paper link: https://www.researchgate.net/publication/305997301_Dynamic_and_Robust_Wildfire_Risk_Prediction_System_An_Unsupervised_Approach
[2] Weather Data: https://www.meteogalicia.gal/observacion/estacionshistorico/consultar.action